@@ -1,21 +1,22 @@
 {
   "name": "moreaboutmohan",
   "homepage": "https://www.moreaboutmohan.com",
-  "version": "2.6.0",
+  "version": "3.0.0",
   "description": "Portfolio - Mohan Subramanian",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.4",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
     "redux": "^4.1.2",
-    "react-redux": "^7.2.6",
+    "react-redux": "^7.2.8",
     "redux-thunk": "^2.4.1",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.5"
   },
   "scripts": {
     "start": "react-scripts start",
