@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { StyledH2 } from 'components/Typography';
 import DesignCardContainer from './styles';
 
-const DesignCard = ({
+function DesignCard({
   defaultImg,
   hoverImg,
   title,
@@ -16,26 +16,28 @@ const DesignCard = ({
   summaryBody,
   clickHandler,
   isCurrent = false,
-}) => (
-  <DesignCardContainer
-    $defaultImg={defaultImg}
-    $hoverImg={hoverImg}
-    className={`design-card  ${isCurrent ? 'selected' : ''}`}
-    onClick={() => clickHandler({
-      pdf,
-      year,
-      time,
-      video,
-      summaryHeading,
-      summaryBody,
-    })}
-  >
-    <span className="design-card__image" />
-    <span className="design-card__overlay" />
-    <StyledH2 className="design-card__title">{title}</StyledH2>
-    <span className="design-card__triangle" />
-  </DesignCardContainer>
-);
+}) {
+  return (
+    <DesignCardContainer
+      $defaultImg={defaultImg}
+      $hoverImg={hoverImg}
+      className={`design-card  ${isCurrent ? 'selected' : ''}`}
+      onClick={() => clickHandler({
+        pdf,
+        year,
+        time,
+        video,
+        summaryHeading,
+        summaryBody,
+      })}
+    >
+      <span className="design-card__image" />
+      <span className="design-card__overlay" />
+      <StyledH2 className="design-card__title">{title}</StyledH2>
+      <span className="design-card__triangle" />
+    </DesignCardContainer>
+  );
+}
 
 DesignCard.propTypes = {
   defaultImg: PropTypes.string,
