@@ -5,15 +5,17 @@ import { StyledH1, StyledH3 } from 'components/Typography';
 import TitleSectionContainer from './styles';
 
 // Component to render the Page title
-const TitleSection = ({
+function TitleSection({
   heading = '',
   subHeading = '',
-}) => (
-  <TitleSectionContainer className="page__title">
-    <StyledH1>{heading}</StyledH1>
-    <StyledH3>{subHeading}</StyledH3>
-  </TitleSectionContainer>
-);
+}) {
+  return (
+    <TitleSectionContainer className="page__title">
+      <StyledH1>{heading}</StyledH1>
+      <StyledH3>{subHeading}</StyledH3>
+    </TitleSectionContainer>
+  );
+}
 
 TitleSection.propTypes = {
   heading: PropTypes.string,
