@@ -13,48 +13,50 @@ import {
 } from './styles';
 
 // Page component for all my photography works and performances
-const PhotographyPage = () => (
-  <PhotographyPageContainer className="photography-page">
-    <Header />
-    <StyledH1>Arts & Photography</StyledH1>
-    <Gradient>
-      <ImageContainer>
-        {
-          imageList.map((image, index) => {
-            if (index % 2 === 1) {
-              return (
-                <LazyLoadImage
-                  key={image}
-                  imageUrl={image}
-                  imageAlt={image}
-                  height="auto"
-                />
-              );
-            }
-            return null;
-          })
-        }
-      </ImageContainer>
-      <ImageContainer>
-        {
-          imageList.map((image, index) => {
-            if (index % 2 === 0) {
-              return (
-                <LazyLoadImage
-                  key={image}
-                  imageUrl={image}
-                  imageAlt={image}
-                  height="auto"
-                />
-              );
-            }
-            return null;
-          })
-        }
-      </ImageContainer>
-    </Gradient>
-    <Footer hoverColor={colors.BLUSH} />
-  </PhotographyPageContainer>
-);
+function PhotographyPage() {
+  return (
+    <PhotographyPageContainer className="photography-page">
+      <Header />
+      <StyledH1>Arts & Photography</StyledH1>
+      <Gradient>
+        <ImageContainer>
+          {
+            imageList.map((image, index) => {
+              if (index % 2 === 1) {
+                return (
+                  <LazyLoadImage
+                    key={image}
+                    imageUrl={image}
+                    imageAlt={image}
+                    height="auto"
+                  />
+                );
+              }
+              return null;
+            })
+          }
+        </ImageContainer>
+        <ImageContainer>
+          {
+            imageList.map((image, index) => {
+              if (index % 2 === 0) {
+                return (
+                  <LazyLoadImage
+                    key={image}
+                    imageUrl={image}
+                    imageAlt={image}
+                    height="auto"
+                  />
+                );
+              }
+              return null;
+            })
+          }
+        </ImageContainer>
+      </Gradient>
+      <Footer hoverColor={colors.BLUSH} />
+    </PhotographyPageContainer>
+  );
+}
 
 export default PhotographyPage;
