@@ -3,7 +3,7 @@ import { Link, useLocation } from 'react-router-dom';
 import { github } from 'images/icons';
 
 // Presentational component for dynamically changing links without resetting the REDUX store
-const Header = () => {
+function Header() {
   const { pathname = '' } = useLocation();
   return (
     <header className="flex-container">
@@ -14,6 +14,6 @@ const Header = () => {
       <a href="https://github.com/maraudermohan/calendar-data"><github.Icon /></a>
     </header>
   );
-};
+}
 
 export default Header;
